using float64 as RequestAnimationFrame callback args

+++ raf/raf.go
@@ -26,10 +26,10 @@ func init() {
 
 	lastTime := 0.0
 	if window.Get("requestAnimationFrame") == js.Undefined {
-		window.Set("requestAnimationFrame", func(callback func(float32)) int {
+		window.Set("requestAnimationFrame", func(callback func(float64)) int {
 			currTime := js.Global.Get("Date").New().Call("getTime").Float()
 			timeToCall := math.Max(0, 16-(currTime-lastTime))
-			id := window.Call("setTimeout", func() { callback(float32(currTime + timeToCall)) }, timeToCall)
+			id := window.Call("setTimeout", func() { callback(float64(currTime + timeToCall)) }, timeToCall)
 			lastTime = currTime + timeToCall
 			return id.Int()
 		})
@@ -42,7 +42,7 @@ func init() {
 	}
 }
 
-func RequestAnimationFrame(callback func(float32)) int {
+func RequestAnimationFrame(callback func(float64)) int {
 	return js.Global.Call("requestAnimationFrame", callback).Int()
 }
